Switch order of ART initation for CD4<200 and >200
According to literature, although persons with greater CD4 counts start
going on ART after 2006, those with lowest CD4 counts are still
prioritized. Therefore, switching the order of modules in hivNH that
calculate persons intiating ART to maintain coverage levels so that
persons with CD4 counts <=200 are priortized over those with >200 cells.

Related to: #67