Fix total pop calibration mean/var & reload .mat
For calibrating to the total population size, used the wrong order of magnitude for the mean and
absolute/standard error (used to calculate variance). Both values were three orders of
magnitude too large. 2011 Census reported value is actually standard error, not
absolute error. Therefore redid calculation of 2011 variance. Had also
not re-saved the calibData.mat file after the recent update of the
variance. Noticed all these issues because the observed mean was >>
model population size, and this was causing a much larger magnitude of
the summed log-likelihood.