Use limits to equalize ART coverage, 2004-2015
Apply new algorithm (which applies minimum and maximum ART coverages by
age while maintaining population-level coverage) beginning with the
start of ART in 2004. In order to do this, had to update helper
functions to take in the relevant HIV disease indices as ART
availability is expanded to persons with increasing CD4 count.

Related to #61