Add calibrated multiplier for male HPV clearance
In calibration round 10Jan20, t=0, the HPV prevalence in men is
consistently too high. Adding a calibrated multiplier that is applied to
female HPV clearance and sets male HPV clearance with range 0.50-0.99.
The same trends by type therefore apply. Literature supports that
clearance may be faster in men.

Related to #79