Decrease HIV/HPV transmission in late-stage HIV/CC
As a proxy for decreased sexual activity due to pain and sickness,
decrease the transmission rate of both HIV and HPV by 50% during
late-stage HIV in men and women and during regional to distant cervical
cancer (stages 2-4) in women. Took this approach because otherwise we
would need to calculate and adjust partnerships and mixing by disease
state. We could do this in the future, but it might require more
substantial model changes and run-time calculations. Would also need to
think more deeply about how this would affect our mixing assumptions.

Streamlined + made changes to HIV code:
- Removed proportion of population practicing anal sex because in our
model it is zero and this is not something we plan to change in the
future.
- The late-stage HIV transmission multiplier was previously set to 0.5.
If we're assuming a 50% decrease in HIV transmission, this should
actually be 7.0*0.5=3.5
- Decrease HIV transmission at all other viral load stages by half with
regional-distant cervical cancer (in late-stage HIV, we've already
accounted for decreased sexual activity).
- Streamline HIV transmission rate matrix by gender
- Update and streamline indices accordingly
- mCurrArt and fCurrArt indices were previously inconsistent. An extra v
ind=1 was set in loadUp2, but not in mixInfect, basically meaning that
the population proportion was summed for only the first index vs. all
appropriate indices. This had no effect because we currently have zero
HIV transmission with ART, but would have mattered more if we assumed
some transmission with ART. This error was introduced with creation of
the nonVaxInfections branch.

Pre-calculated HPV transmission values where possible + made changes to
HPV code:
- Formatted similarly to HIV code
- Updated indices accordingly
- Fixed calculation of HPV beta by age and risk. For a person of age i
and risk j partnering with a proportion of the population of age ii and
risk jj, the correct HPV transmission rate is a function of the number
of acts the person of age i and risk j has. Previously this was always a
function of the acts of the partner of age ii and risk jj which is
incorrect. The result of this fix is less HPV in the youngest and oldest
ages and higher HPV prevalence in the middle ages.

Fixes #58
Related to #62
Related to #49