Decrease HIV/HPV transmission for late-stage HIV
Decrease HIV and HPV transmission by 90% as a proxy for decreased sexual
activity and sickness during late-stage HIV. Setting the relative risk
for late-stage HIV to be 0.7 (0.10*7) so the reduction in transmission
is a more rational number. In other words, framing the reduction in risk
relative to what it would be in the AIDS stage, making the assumption
that it would be 90% lower, and 0.7 times that in the asymptomatic
stage. Likewise reduce transmission of HPV for persons with late-stage
HIV by 90%.

Don't have much empirical data on this either, but going to continue to
assume 50% reduction in HIV/HPV transmission with cervical cancer.
People live years with stage 2+ cervical cancer rather than only 8
months with late-stage HIV. So the health stages are not comparable in
terms of assumption of how sick and close to death people are.

If a person has late-stage HIV and stage 2+ cervical cancer, assume the
lowest sexual activity, ie. 90% HPV/HIV transmission reduction.

Related to #58
Related to #62