Plot prop of all HIV+ on ART vs. updated data
Add plot of ART coverage relative to all HIV-positive persons
rather than only persons with an eligible CD4 count. Update observed
data we compare to to capture the proportion of persons on ART who are
virally suppressed and the most recent data reviewed in literature. Add
code to feed in these values.

Related to #67