Enable auto-updating & increase initial population size
Change assumption about the historical decline of the KZN
provincial share of the SA population. Increased the slope to assume a
larger initial provincial share in 1950, which is used to estimate the
KZN population size in 1950. The initial population size in 1925 is then
back-extrapolated.

Use equations rather than built-in Excel trendlines to calculate
exponential trend for KZN population changes by age from 1925-1950. This
enables auto-updating of estimated 1925 population size by gender and
age with changes to assumptions about the KZN provincial share in 1950.