Decrease year ART reaches 90-90-90 target to 2029 vs 2030
ART coverage scale-up lags in the model compared to observed data.
I think this is because the model is equilibrating and the CD4
threshold is changing. Therefore, I moved coverage scale-up to
one year prior each benchmark before 2020. However, forgot to do
this in the hivNH function for year 2030 as well (the year we
assume KZN reaches 90-90-90). Was therefore getting an error
because 2030 wasn't defined in the coverage vector. This year
should instead be defined in the code as 2030-1.

Related to #67