Remove modifications to gender-age bias in mixing by age matrices
I noticed some modifications were made to the assortative mixing
matrix by age that I think are a little funky. These modifications
remove the gender-age bias in persons <=19, and make it so 20-24
year-old males instead of 10-14 year-old males have less than 100%
of their partnerships. I think 20-24 year olds are more influential
to transmission in the model so we should remove these modifications.

Also set proportion partnerships to zero for those age groups that
are not sexually active.

Related to #49