Remove extra mult on CIN2->CIN1 regression in HIV+ women
The multiplier on CIN2->CIN1 regression in HIV-positive women
was set to be the square of the multiplier used on HPV clearance
and CIN3->CIN2 regression in HIV-positive women. In the absense of
evidence that this should be the case, removed this extra compounding
multiplier. Now there is a consistent multiplier for HPV clearance
regression for women living with HIV.

Related to #34