Adjust baseline acts/partnership/year for both genders
Change baseline values when not calibrating (this is not
currently a calibrated parameter). Was seeing a high peak
in female HPV prevalence at ages 25-29 due to a really high
number of acts and low prevalence in older ages due to very
few acts. The below changes allow us to better fit the
female HPV prevalence curve by age.
-Adjust female acts, increasing acts in younger ages and
decreasing acts in middle ages
-Assume a constant number of acts/partnership/year for ages
45+ vs. a decline in acts with age
-Set male acts based on female acts, skewing to be relatively
lower in younger ages and relatively higher in middle ages to
reflect bias in partnerships
-Set male acts to be the same as females in ages 50-79

Related to #73