Change mults on 10-24 HPV/CIN/CC transition rates for older ages
Observed data for Cape Town, SA shows a decline in HPV prevalence
with increasing age. To decrease prevalence in middle ages and
achieve this shape (and also just decrease HPV prevalence in
general, which tended to be too high in my hand calibration),
I changed the multiplier for 25-49 year-olds on the HPV
clearance rate for 10-24 year-olds to 1.0. This effectively
increased clearance in this age group and decreased HPV prevalence.
I removed changes to the multipliers for older ages to minimize
the number of parameters I was adjusting.

To increase progression to CIN, I changed the multipliers for older
ages on the HPV to CIN1 progression rate for ages 10-24. Unless
we find additional observed data, it didn't make sense that
progression at other lesion/cancer stages would increase with age
but that HPV to CIN1 progression would decrease with age. This
also allowed me to increase the clearance rate from HPV and the
regression rate from CIN1 to HPV to better fit female HPV prevalence
while maintaining reasonable HPV transmission and maleHpvClearMult
values.. The new multipliers are 1.05, 1.10, and 1.15 for ages 25-49,
50-69, and 70-79, respectively.