Increase multipliers on CIN progression in WLWHIV
Slightly increase the multipliers on CIN1 to CIN2 and
CIN2 to CIN3 progression in women living with HIV with
CD4 <500. Did this to increase CIN prevalence in older
ages among women living with HIV (had a small effect,
but challenging to target age groups by adjusting CD4-
correlated multipliers).

Related to #34