Apply baseline screening coverage to full age group
Previously multiplied coverage by 0.2 to target a single age.
However, this under-estimates screening coverage, because women
aging out of the 5-year age group have coverage 1/5 of the target
coverage. Instead applying the full coverage to the entire
compartment. This results in a higher number of women getting
screened each time target coverage is increased, but since women
are not repeatedly screened in their time that compartment, the
effect diminishes. Therefore, we are in effect targeting one-time
screening to 35-39 year-olds.

Related to #82