Decrease alpha proportion accepted particles to 0.4
Alpha was set at 0.6. Decreasing to 0.4 to still be in the recommended
range from (Lenormand, 2012) but because it is suggested that smaller
values are alpha are slightly better for larger values of p_acc. Hoping
this will improve the quality of the estimated posterior without a huge
computational impact due to parallelization on the checkpoint queue.