Update historical VMMC & ART for DO ART scenarios
Continued ramp-up of VMMC from 2017 to 2020 at 2012 to 2017 rate.
This represents the baseline DO ART VMMC scenario.

Ramp-up to DO ART clinic ART levels in females in addition to males
in 2017. Carry these values forward to 2020. This represents less
ART in males and more ART in females compared to the values used
during model calibration.